Stricter separation of nodal and modal discretization routines
